+++ CHANGELOG.md
@@ -2,6 +2,19 @@
 
 All notable changes to this project will be documented in this file.
 
+## [v0.2.9] - 2026-04-27
+
+
+
+### Changed
+
+- Add THIRD_PARTY_LICENSE.txt with licenses for included dependencies
+
+- Rename build:win script and update versioning process to copy LICENSE file
+
+- Remove THIRD_PARTY_LICENSE.txt as it is no longer needed
+
+
 ## [v0.2.8] - 2026-04-27